Add empty function definition for unit test
Previously the functions called in automationHelper were defined outside
the scope of this module, but were already present on the stemcell as
these calls were backfilling existing functionality. For a new function
such as Set-InternetExplorerRegistries that is not yet defined in a
stemcell we explictly defined a blank implementation in the test so the
unit tests can pass without relying on external dependencies.

[#167407675](https://www.pivotaltracker.com/story/show/167407675)
 As a platform engineer I want to see Internet Explorer-based policies that comply with MSFT Baseline Security Standard - 2019

Co-authored-by: Daniel Basner <dbasner@pivotal.io>